requirements.txt
selenium>=4.27.0
webdriver-manager>=4.0.0
gspread>=6.2.1
google-auth-oauthlib>=1.2.4
google-api-python-client>=2.190.0
streamlit>=1.54.0
python-dotenv>=1.1.0
pyautogui>=0.9.54
